+++ src/PostgREST/Config/Database.hs
@@ -61,7 +61,7 @@ dbSettingsNames =
   ,"jwt_role_claim_key"
   ,"jwt_secret"
   ,"jwt_secret_is_base64"
-  ,"jwt_cache_max_lifetime"
+  ,"jwt_cache_max_entries"
   ,"openapi_mode"
   ,"openapi_security_active"
   ,"openapi_server_proxy_uri"